feat(cb2-10339): Addition of IVA Defect validation for PUT endpoint (#374)
* feat(cb2-10339): added ivaDefet Joi validation to the PUT endpoint

* feat(cb2-10339): added iva defects joi validation to non-specialist test types
