+++ src/models/validators/SpecialistTestsCommonSchemaSubmitted.ts
@@ -1,46 +1,10 @@
 import * as Joi from 'joi';
-import { ValidationErrorItem } from 'joi';
 import {
   defectsCommonSchema,
+  ivaDefectSchema,
   testResultsCommonSchema,
   testTypesCommonSchema,
 } from './CommonSchema';
-
-const customInspectionTypesErrorMessage = (
-  errors: ValidationErrorItem[],
-): ValidationErrorItem[] =>
-  errors.map((e) =>
-    e.type === 'any.allowOnly'
-      ? { ...e, message: `"inspectionTypes" must be one of [basic, normal]` }
-      : e,
-  );
-
-export const ivaDefectSchema = Joi.object({
-  sectionNumber: Joi.string().required(),
-  sectionDescription: Joi.string().required(),
-  additionalInformation: Joi.object({
-    notes: Joi.string().required(),
-  }).optional(),
-  requiredStandards: Joi.array()
-    .items(
-      Joi.object({
-        rsNumber: Joi.number().required(),
-        requiredStandard: Joi.string().required(),
-        refCalculation: Joi.string().required(),
-        additionalInfo: Joi.boolean().required(),
-        inspectionTypes: Joi.array()
-          .items(
-            Joi.string()
-              .valid('basic', 'normal')
-              .error(customInspectionTypesErrorMessage),
-          )
-          .min(1)
-          .max(2)
-          .required(),
-      }),
-    )
-    .required(),
-});
 
 export const testTypesIVADefectCommonSchemaSpecialistTestsSubmitted =
   testTypesCommonSchema.keys({
