+++ src/models/validators/TestResultsSchemaMotorcycleSubmitted.ts
@@ -1,10 +1,10 @@
 import { any, array, object } from 'joi';
 import {
   defectsCommonSchemaSpecialistTestsSubmitted,
-  ivaDefectSchema,
   testResultsCommonSchemaSpecialistTestsSubmitted,
   testTypesCommonSchemaSpecialistTestsSubmitted,
 } from './SpecialistTestsCommonSchemaSubmitted';
+import { ivaDefectSchema } from './CommonSchema';
 
 export const motorcycleSubmitted =
   testResultsCommonSchemaSpecialistTestsSubmitted.keys({
